Create a first concept for the UI
Alas, I want to make this in Godot, but making SQLite work with Godot 4
is impossible at the moment.  Available extensions don't yet support
Godot 4 and interfacing with the browser to use SQL.js is impossible due
to limitations in Godot's JS FFI implementation.

But hey, it works!  And I was able to do some basic tests on Linux
Firefox and Chromium and Android Firefox, Chromium, and Samsung Browser
to validate that everything works on most browsers.