Rework the Question-Answer form to be clearer
When I showed Oliver (native speaker) this site, he didn't at first
recognize that the big word in bold at the top was the word he should be
conjugating.  That meant that the UI's structure didn't make that clear.

Lining up the left edges of the prompt and answer I feel makes it clear
they're related.  As a bonus, it helps to diagnose errors due to bad
typing since the letters line up (though you need to explicitly make the
prompt and answer be the same font for this to work).