Reflow the UI to make the point of the site clear
When observing people trying out the site, usually people are a bit
confused at first and I need to explain to them what to do.  One person
even thought it was a search tool instead of a game.

I've now made the prompt and the answer be horizontal to each other
instead of vertical and placed an arrow between so players realize one
leads to the other.

There's also a big h1 heading at the top now so people don't think that
the prompt is the page title or something.