Center the Arvaa button below the answer/prompt
Everyone I show this to initially thinks this is for looking up word
forms, including Finnish speakers who can natively read the UI text.  My
latest theory is that it's because the Arvaa (guess) button is next to
the input box, which is similar to how search is typically drawn.

I'm guessing though that I'll need to add some colors and flair so it
looks more gamey to get people into the gaming mood.