Add a noun table
I'd like to be able to practice noun inflections, so there needs to be
that data available.  It was relatively easy to assemble the table.
Basically two evenings mostly copying the verb table generation. Finding
each form with this loop and match technique is a huge performance boost
over what the verb generation does, so I'll have to implement that there
later.

The generated table is 45 MB vs 4.1 MB for the verb table, so I may need
to think about culling words.

Next step is to make a UI!