Refactor verb parsing to be like noun parsing
I was excited by the potential performance boost and was not
disappointed.  1.929 ms/word before, 0.491 ms/word after.

It's actually quite a lot longer than for nouns.  I wonder if it's
because I'm getting the gradation and conjugation, or if it's because
the sets of tags that it's matching against is longer.